feat: Enhance car details, add accessibility features, and expand test coverage
This commit updates the data model to support detailed vehicle attributes, improves the UI with accessibility options, and introduces a comprehensive testing suite.

- **Data Layer:**
    - Update `Car` entity to include `combustivel`, `categoria`, and `acessorios` fields.
    - Add Room `TypeConverters` to handle list serialization for accessories.
    - Change `carId` type from `Int` to `Long` across `CarDao`, `Poster` entity, and repository calls.
    - Increment `AppDatabase` version to 10.

- **UI and Accessibility:**
    - Implement `AboutScreen` displaying app info and an email intent for support.
    - Add font scaling controls (A+/A-) and branding logos to `LoginScreen`, `RegisterScreen`, and `CarDetailScreen`.
    - Refactor the advertisement creation flow in `AnunciosScreen` into `CustomCarDialog`, featuring dropdowns for category/fuel and checkboxes for accessories.
    - Update `CarDetailScreen` to visualize new attributes using a `Badge` component and `FlowRow`.

- **Testing and Config:**
    - Add unit tests for business logic: `LoginUnitTest`, `RegisterLogicTest`, `CarDetailUnitTest`, and `AnunciosUnitTest`.
    - Implement UI instrumentation tests for `LoginScreen`.
    - Update `build.gradle.kts` and `libs.versions.toml` with KSP configuration and testing dependencies.
    - Remove the default `ExampleUnitTest`.

+++ app/src/main/java/com/example/luxcar/data/model/Poster.kt
@@ -11,5 +11,5 @@ data class Poster(
     val descricao: String,
     val preco: Double,
     val imagem: ByteArray,
-    val carId: Int
+    val carId: Long
 )
